+++ clients/bendini/bendini.nix
@@ -12,11 +12,11 @@ base.languages.rust.mkClient rec {
 
   shellHook = ''
     bendiniLocal() {
-      command cargo run -- --host "$XDG_RUNTIME_DIR"/avery.sock --auth-host "$XDG_RUNTIME_DIR"/avery.sock "$@"
+      command cargo run --quiet -- --host unix://localhost"$XDG_RUNTIME_DIR"/avery.sock --auth-host unix://localhost"$XDG_RUNTIME_DIR"/avery.sock "$@"
     }
 
     bendiniLocalAuth() {
-      command cargo run -- --auth-host "$XDG_RUNTIME_DIR"/avery.sock "$@"
+      command cargo run --quiet -- --auth-host unix://localhost"$XDG_RUNTIME_DIR"/avery.sock "$@"
     }
 
     echo -e "Use \e[32;1mbendiniLocal\e[0m to run bendini against a development version of avery \e[36;1m(both auth and functions)\e[0m"