+++ protocols/CHANGELOG.md
@@ -6,7 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Name filter is now just a string instead of a type
+
 ### Added
+- ListVersions endpoint for registry.
+- publisher_email field for Filters in registry listings.
 - Publisher field to FunctionData, Function, AttachmentData and Attachment.
 - GetIdentity endpoint to auth.