+++ services/avery/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Implemented endpoint for GetIdentity
+- Implemented ListVersions endpoint
+- Added publisher email filter for List and ListVersion
 
 ## [1.2.1] - 2021-10-21